Safari-609.1.11.1/LayoutTests/svg/animations/animate-from-to-keyTimes-expected.txt

SVG 1.1 dynamic animation tests

Tests discrete from-to-keyTimes animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 200
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 200
PASS rect.x.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
